chore: refactor using clsx
