+++ packages/content/src/graphql/mutations/AddResourceToQueue.graphql
@@ -1,0 +1,3 @@
+mutation AddResourceToQueue($resourceType: ContentKind, $resourceId: ID!) {
+  AddResourceToQueue(resourceType: $resourceType, resourceId: $resourceId)
+}
